Remove Methode preview
