--- helm/content-unroller/app-configs/content-preview-unroller_eks_delivery.yaml
@@ -1,18 +1,0 @@
-# Values used for the deployed application.
-replicaCount: 2
-service:
-  name: content-preview-unroller
-eksCluster: true
-
-content_store:
-  app_name: content-public-read
-  host: http://content-public-read:8080
-
-content_preview:
-  app_name: content-public-read-preview
-  host: http://content-public-read-preview:8080
-
-content_path: /content
-internal_content_path: /internalcontent
-
-flow: preview
