+++ helm/content-unroller/app-configs/content-unroller_delivery.yaml
@@ -13,5 +13,3 @@ content_preview:
 
 content_path: /content
 internal_content_path: /internalcontent
-
-flow: read
