+++ helm/content-unroller/app-configs/content-unroller_eks_delivery.yaml
@@ -14,5 +14,3 @@ content_preview:
 
 content_path: /content
 internal_content_path: /internalcontent
-
-flow: read
