--- runbooks/content-preview-unroller_runbook.md
@@ -1,126 +1,0 @@
-# UPP - Content Preview Unroller
-
-This service (internal API) expands images and dynamic content of an article. There are two deployments of the same code base - this one is part of the preview flow, the other one (content-unroller) is part of the reading flow.
-
-## Code
-
-content-preview-unroller
-
-## Primary URL
-
-<https://upp-prod-delivery-glb.upp.ft.com/__content-unroller/>
-
-## Service Tier
-
-Bronze
-
-## Lifecycle Stage
-
-Production
-
-## Delivered By
-
-content
-
-## Supported By
-
-content
-
-## Known About By
-
-- dimitar.terziev
-- hristo.georgiev
-- elitsa.pavlova
-- elina.kaneva
-- kalin.arsov
-- ivan.nikolov
-- miroslav.gatsanoga
-- mihail.mihaylov
-- tsvetan.dimitrov
-- georgi.ivanov
-- robert.marinov
-
-## Host Platform
-
-AWS
-
-## Architecture
-
-Content Preview Unroller is an internally used API for expanding images and dynamic content of an article preview. It receives an article and returns the same article with some additional data: main image, body embedded images, alternative images, lead images and embedded dynamic content.
-
-## Contains Personal Data
-
-No
-
-## Contains Sensitive Data
-
-No
-
-## Dependencies
-
-- content-public-read-preview
-
-## Failover Architecture Type
-
-ActiveActive
-
-## Failover Process Type
-
-FullyAutomated
-
-## Failback Process Type
-
-FullyAutomated
-
-## Failover Details
-
-The service is deployed in both Delivery clusters.
-The failover guide for the cluster is located here:
-<https://github.com/Financial-Times/upp-docs/tree/master/failover-guides/delivery-cluster>
-
-## Data Recovery Process Type
-
-NotApplicable
-
-## Data Recovery Details
-
-The service does not store data, so it does not require any data recovery steps.
-
-## Release Process Type
-
-PartiallyAutomated
-
-## Rollback Process Type
-
-Manual
-
-## Release Details
-
-Manual failover is needed when a new version of
-the service is deployed to production.
-Otherwise, an automated failover is going to take place when releasing.
-For more details about the failover process please see: <https://github.com/Financial-Times/upp-docs/tree/master/failover-guides/delivery-cluster>
-
-## Key Management Process Type
-
-Manual
-
-## Key Management Details
-
-To access the service clients need to provide basic auth credentials.
-To rotate credentials you need to login to a particular cluster and update varnish-auth secrets.
-
-## Monitoring
-
-Service in UPP K8S delivery clusters:
-
-- Delivery-Prod-EU health: <https://upp-prod-delivery-eu.ft.com/__health/__pods-health?service-name=content-unroller>
-- Delivery-Prod-US health: <https://upp-prod-delivery-us.ft.com/__health/__pods-health?service-name=content-unroller>
-
-## First Line Troubleshooting
-
-<https://github.com/Financial-Times/upp-docs/tree/master/guides/ops/first-line-troubleshooting>
-
-## Second Line Troubleshooting
-
-Please refer to the GitHub repository README for troubleshooting information.
